+++ src/Game/Client/Physics/Physics.cpp
@@ -1132,7 +1132,7 @@ void CLG_Physics_Step(SGEntityHandle &entityHandle)
 
     // Friction for flying monsters that have been given vertical velocity
     if ((ent->GetFlags() & EntityFlags::Fly) && (ent->GetVelocity().z != 0)) {
-        float speed = std::fabsf(ent->GetVelocity().z);
+        float speed = fabs(ent->GetVelocity().z);
         float control = speed < STEPMOVE_STOPSPEED ? STEPMOVE_STOPSPEED : speed;
         float friction = STEPMOVE_FRICTION / 3;
         float newSpeed = speed - (CLG_FRAMETIME * control * friction);
@@ -1145,7 +1145,7 @@ void CLG_Physics_Step(SGEntityHandle &entityHandle)
 
     // Friction for flying monsters that have been given vertical velocity
     if ((ent->GetFlags() & EntityFlags::Swim) && (ent->GetVelocity().z != 0)) {
-        float speed = std::fabsf(ent->GetVelocity().z);
+        float speed = fabs(ent->GetVelocity().z);
         float control = speed < STEPMOVE_STOPSPEED ? STEPMOVE_STOPSPEED : speed;
         float newSpeed = speed - (CLG_FRAMETIME * control * STEPMOVE_WATERFRICTION * ent->GetWaterLevel());
         if (newSpeed < 0)
